add turret stats
