+++ Source/RangedDPS.csproj
@@ -47,6 +47,8 @@
     <Compile Include="StatWorker_RangedDPSBase.cs" />
     <Compile Include="StatWorker_MaxRangedDPS.cs" />
     <Compile Include="StatWorker_TurretRangedDPS.cs" />
+    <Compile Include="StatWorker_TurretRangedDPSBase.cs" />
+    <Compile Include="StatWorker_MaxTurretRangedDPS.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
