+++ Source/StatWorker_TurretRangedDPS.cs
@@ -1,21 +1,14 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
+using System.Text;
 using RimWorld;
 using Verse;
 
 namespace RangedDPS
 {
 
-    public class StatWorker_TurretRangedDPS : StatWorker_RangedDPSBase
+    public class StatWorker_TurretRangedDPS : StatWorker_TurretRangedDPSBase
     {
-        public override bool ShouldShowFor(StatRequest req)
-        {
-            if (!(req.Def is ThingDef thingDef))
-            {
-                return false;
-            }
-
-            return thingDef?.building?.IsTurret ?? false;
-        }
 
         public override float GetValueUnfinalized(StatRequest req, bool applyPostProcess = true)
         {
@@ -26,6 +19,10 @@ namespace RangedDPS
             }
 
             var weaponDef = thingDef?.building?.turretGunDef;
+            if (weaponDef == null)
+            {
+                return 0f;
+            }
 
             var shootVerb = GetShootVerb(weaponDef);
             if (shootVerb == null)
@@ -45,54 +42,60 @@ namespace RangedDPS
             return rawDps * bestAccuracy;
         }
 
-        public override string GetExplanationUnfinalized(StatRequest req, ToStringNumberSense numberSense)
-        {
-            ThingDef thingDef = req.Def as ThingDef;
-            if (thingDef == null)
-            {
-                return "";
-            }
+        //TODO make this work given that turret guns have no Thing
+        //public override string GetExplanationUnfinalized(StatRequest req, ToStringNumberSense numberSense)
+        //{
+        //    ThingDef thingDef = req.Def as ThingDef;
+        //    if (thingDef == null)
+        //    {
+        //        return "";
+        //    }
 
-            var weaponDef = GetWeaponDef(thingDef);
-            var shootVerb = GetShootVerb(weaponDef);
-            if (shootVerb == null)
-            {
-                return "";
-            }
+        //    var weaponDef = thingDef?.building?.turretGunDef;
+        //    if (weaponDef == null)
+        //    {
+        //        return "";
+        //    }
 
-            float rawDps = GetRawDPS(shootVerb, weaponDef);
+        //    var shootVerb = GetShootVerb(weaponDef);
+        //    if (shootVerb == null)
+        //    {
+        //        return "";
+        //    }
 
-            StringBuilder stringBuilder = new StringBuilder();
-            stringBuilder.AppendLine("StatsReport_RangedDPSAccuracy".Translate());
+        //    float rawDps = GetTurretRawDPS(shootVerb, weaponDef);
 
-            for (int i = 5; i <= shootVerb.range; i += 5)
-            {
-                float hitChance = shootVerb.GetHitChanceFactor(req.Thing, i);
-                float damage = rawDps * hitChance;
-                stringBuilder.AppendLine(string.Format("{0} {1,2}: {2,5:F2} ({3:P1})",
-                    "distance".Translate().CapitalizeFirst(),
-                    i, damage, hitChance));
-            }
+        //    StringBuilder stringBuilder = new StringBuilder();
+        //    stringBuilder.AppendLine("StatsReport_RangedDPSAccuracy".Translate());
 
-            // Max Range
-            float maxRangeHitChance = shootVerb.GetHitChanceFactor(req.Thing, shootVerb.range);
-            float maxRangeDamage = rawDps * maxRangeHitChance;
-            stringBuilder.AppendLine(string.Format("{0} {1,2}: {2,5:F2} ({3:P1})",
-                    "distance".Translate().CapitalizeFirst(),
-                    shootVerb.range, maxRangeDamage, maxRangeHitChance));
+        //    int startRange = 5;
+        //    if (shootVerb.minRange > 0)
+        //    {
+        //        float minRangeHitChance = shootVerb.GetHitChanceFactor(req.Thing, shootVerb.minRange);
+        //        float minRangeDps = rawDps * minRangeHitChance;
+        //        stringBuilder.AppendLine(string.Format("{0} {1,2}: {2,5:F2} ({3:P1})",
+        //                "distance".Translate().CapitalizeFirst(),
+        //                shootVerb.minRange, minRangeDps, minRangeHitChance));
 
-            return stringBuilder.ToString();
-        }
+        //        int startPoint = (int)Math.Ceiling(shootVerb.minRange / 5) * 5;
+        //    }
 
-        protected float GetTurretRawDPS(VerbProperties shootVerb, ThingDef weaponDef)
-        {
-            float fullCycleTime = shootVerb.warmupTime + weaponDef.GetStatValueAbstract(StatDefOf.RangedWeapon_Cooldown)
-                    + ((shootVerb.burstShotCount - 1) * shootVerb.ticksBetweenBurstShots).TicksToSeconds();
+        //    for (int i = startRange; i <= shootVerb.range; i += 5)
+        //    {
+        //        float hitChance = shootVerb.GetHitChanceFactor(req.Thing, i);
+        //        float damage = rawDps * hitChance;
+        //        stringBuilder.AppendLine();
+        //    }
 
-            int totalDamage = shootVerb.burstShotCount * shootVerb.defaultProjectile.projectile.GetDamageAmount_NewTmp(weaponDef, null);
+        //    // Max Range
+        //    float maxRangeHitChance = shootVerb.GetHitChanceFactor(req.Thing, shootVerb.range);
+        //    float maxRangeDps = rawDps * maxRangeHitChance;
+        //    stringBuilder.AppendLine(string.Format("{0} {1,2}: {2,5:F2} ({3:P1})",
+        //            "distance".Translate().CapitalizeFirst(),
+        //            shootVerb.range, maxRangeDps, maxRangeHitChance));
 
-            return totalDamage / fullCycleTime;
-        }
+        //    return stringBuilder.ToString();
+        //}
 
     }
 }
